requirements.txt
ccxt>=4.0.0
python-dotenv>=1.0.0
aiohttp>=3.9.0
numpy>=1.24.0
pandas>=2.0.0
gymnasium>=0.29.0
stable-baselines3>=2.0.0
torch>=2.0.0
pybit>=5.6.2
python-telegram-bot>=20.7
